backend/requirements.txt
flask==2.1.0
flask-sqlalchemy==2.5.1
flask-marshmallow==0.14.0
flask-bcrypt==0.7.1
flask-jwt-extended==4.4.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
webrtc-python==1.0.2
email-validator==1.1.3
gunicorn==20.1.0
flask-cors==3.0.10
